adding authetication on each routes and making some modificatios on API endpoints

+++ server/controllers/bookingController.js
@@ -1,9 +1,9 @@
-const { QueryTypes } = require("sequelize");
+const { QueryTypes, json } = require("sequelize");
 const sequelize = require("../config/db");
 const Booking = require("../models/booking")
 const BookingDetails = require("../models/bookingDetails")
 const BookingHistory = require("../models/bookingHistory");
-const room = require("../models/room");
+const Room = require("../models/room");
 
 //Joins
 const sqlBooking =`
@@ -36,6 +36,25 @@ LEFT JOIN promos_and_offers p ON bd.offer_id = p.offer_id
 const sqlBookingGroupBy = `
  GROUP BY b.booking_id, b.check_in_date, b.check_out_date, b.total_amount, c.customer_id, c.first_name, c.middle_name, c.last_name, c.email
 `
+
+//using a helper function to check room availablilty for a specified date range so that booking wont overlap
+async function isRoomAvailable(room_no,check_in_date,check_out_date){
+    const res = await sequelize.query(
+        `
+        SELECT 1 FROM booking b
+        JOIN booking_details bd ON b.booking_id=bd.booking_id
+        WHERE bd.room_no = :room_no
+        AND NOT(b.check_out_date <= :check_in_date OR b.check_in_date >= :check_out_date)
+        LIMIT 1
+        `,
+        {
+            replacements: {room_no, check_in_date, check_out_date},
+            type:QueryTypes.SELECT
+        }
+    )
+    return !res || res.length === 0 
+}
+
 //Booking Creation
 const createBooking = async (req,res) => {
     try{
@@ -47,21 +66,28 @@ const createBooking = async (req,res) => {
         (new Date(check_out_date)- new Date(check_in_date))/(1000*60*60*24)
 )
 
-        for (const room of rooms){
+        for (const rm of rooms){
+            //check availablilty
+            const available = await isRoomAvailable(rm.room_no, check_in_date, check_out_date)
+            if(!available)
+                return res.status(400).json({message: `Room ${rm.room_no} is already booked`})
             const [roomData] = await sequelize.query(
-                `SELECT price_per_night FROM room WHERE room_no= :room_no`,
+                `SELECT price_per_night FROM room WHERE room_no= :room_no AND room_status = '1'`,
                 {
-                    replacements: {room_no: room.room_no},
+                    replacements: {room_no: rm.room_no},
                     type: QueryTypes.SELECT
                 }
             )
+            if(!roomData)
+                return res.status(400).json({message: `Room ${rm.room_no} is not available`})
+            
             let roomPrice = roomData.price_per_night; 
-                //for offers
-            if(room.offer_id){
+            //for offers
+            if(rm.offer_id){
                 const [offerData] = await sequelize.query(
                     `SELECT offered_discount FROM promos_and_offers WHERE offer_id= :id`,
                     {
-                        replacements: {id:room.offer_id},
+                        replacements: {id:rm.offer_id},
                         type:QueryTypes.SELECT
                     }
                 )
@@ -74,8 +100,8 @@ const createBooking = async (req,res) => {
 
             //to keep track of details
             roomDetails.push({
-                room_no: room.room_no,
-                offer_id:room.offer_id || null,
+                room_no: rm.room_no,
+                offer_id:rm.offer_id || null,
                 price_per_night: roomPrice
             })
         }
@@ -100,40 +126,15 @@ const createBooking = async (req,res) => {
                 offer_id:detail.offer_id
             })
 
-            await room.update(
+            await Room.update(
                 {room_status: "0"},
                 {where: {room_no: detail.room_no}}
             )
         }
-        //     await sequelize.query(
-        //         `INSERT INTO booking_details (booking_id, room_no, offer_id)
-        //         VALUES(:booking_id, :room_no, :offer_id)`,
-        //         {
-        //             replacements: {
-        //                 booking_id: booking.booking_id,
-        //                 room_no: detail.room_no,
-        //                 offer_id: detail.offer_id || null,
-        //             },
-        //             type: QueryTypes.INSERT
-        //         }
-        //     )
-        //}
-
         await BookingHistory.create({
             customer_id,
             booking_id: booking.booking_id
         })
-        // await sequelize.query(
-        //     `INSERT INTO booking_history(customer_id, booking_id)
-        //     VALUES(:customer_id, :booking_id)`,
-        //     {
-        //         replacements: {
-        //             customer_id,
-        //             booking_id:booking.booking_id
-        //         },
-        //         type: QueryTypes.INSERT
-        //     }
-        // )
         res.status(201).json(booking)
     }catch(err){
         res.status(500).json({error: err.message});
@@ -143,7 +144,9 @@ const createBooking = async (req,res) => {
 //GET ALL booking
 const getAllBooking = async (req,res) => {
     try{
-        const booking = await sequelize.query(`${sqlBooking} ${sqlBookingGroupBy}`);
+        const booking = await sequelize.query(`${sqlBooking} ${sqlBookingGroupBy}`,{
+            type:QueryTypes.SELECT
+        });
         res.json(booking)
     }catch(err){
         res.status(500).json({error: err.message})
@@ -180,8 +183,16 @@ const updateBooking = async(req,res) => {
         if(!booking)
             return res.status(404).json({message: 'Booking not found'})
 
+        //only the admin or the cusotmer that booked can perform updation
+        if(req.user.role !== "admin" && req.user.id !== booking.customer_id)
+            return res.status(401).json({message:"Unauthorized:Access Denied"})
+        
+        //preventing update after check-in
+        if(new Date(booking.check_in_date)<= new Date())
+            return res.status(400).json({error: "Cannot modify past bookings"})
+
         await booking.update(req.body)
-        res.json(booking)
+        res.json({message: "booking updated successfully",booking})
     }catch(err){
         res.status(500).json({error: err.message})
     }
@@ -191,10 +202,24 @@ const updateBooking = async(req,res) => {
 const deleteBooking = async (req,res) => {
     try{
         const booking = await Booking.findByPk(req.params.id)
+        const bookedRooms = await BookingDetails.findAll({where: { booking_id: req.params.id }})
         if(!booking)
             return res.status(404).json({message: 'Booking not found'})
+
+        //only admin or customer that booked can perform deletion
+        if(req.user.role !== "admin" && req.user.id !== booking.customer_id)
+            return res.status(401).json({message:"Unauthorized:Access Denied"})
+
+        //preventing cancellation after check-in
+        if(new Date(booking.check_in_date)<= new Date())
+            return res.status(400).json({error: "Cannot delete past bookings"})
+        
         await BookingDetails.destroy({ where: { booking_id: req.params.id } });
         await BookingHistory.destroy({ where: { booking_id: req.params.id } });
+        
+        for(const r of bookedRooms){
+            await Room.update({room_status: '1'},{where: {room_no: r.room_no}})
+        }
         await booking.destroy();
         res.json({message: 'Booking deleted successfully'})
     }catch(err){
@@ -229,7 +254,7 @@ const searchBookingByCDetail = async (req,res) => {
     res.json(booking)
     }
     catch(err){
-        res.status(500).json*{error: err.message}
+        res.status(500).json({error: err.message})
     }   
 }
 exports.createBooking = createBooking
